Pull paths.ts out from config.js
